tests

+++ tests/unittests/datamodel/test_gt_colorproperty.cpp
@@ -72,15 +72,15 @@ TEST_F(TestGtColorProperty, toRGB)
     m_prop->setVal(QColor(Qt::blue).name());
     EXPECT_STREQ(m_prop->hexString().toStdString().c_str(), "#0000FF");
     gt::rgb test1 = m_prop->toRGB();
-    EXPECT_EQ(test1.m_r, 0);
-    EXPECT_EQ(test1.m_g, 0);
-    EXPECT_EQ(test1.m_b, 255);
+    EXPECT_EQ(test1.m_r, uint8_t(0));
+    EXPECT_EQ(test1.m_g, uint8_t(0));
+    EXPECT_EQ(test1.m_b, uint8_t(255));
 
     m_prop->setVal(QColor(Qt::yellow).name());
     gt::rgb test2 = m_prop->toRGB();
-    EXPECT_EQ(test2.m_r, 255);
-    EXPECT_EQ(test2.m_g, 255);
-    EXPECT_EQ(test2.m_b, 0);
+    EXPECT_EQ(test2.m_r, uint8_t(255));
+    EXPECT_EQ(test2.m_g, uint8_t(255));
+    EXPECT_EQ(test2.m_b, uint8_t(0));
 }
 
 TEST_F(TestGtColorProperty, toRGBandAlpha)
@@ -88,17 +88,17 @@ TEST_F(TestGtColorProperty, toRGBandAlpha)
     m_prop->setFromRGB(gt::rgb(0, 0, 255, 255));
     EXPECT_STREQ(m_prop->hexString().toStdString().c_str(), "#0000FF");
     gt::rgb test1 = m_prop->toRGB();
-    EXPECT_EQ(test1.m_r, 0);
-    EXPECT_EQ(test1.m_g, 0);
-    EXPECT_EQ(test1.m_b, 255);
-    EXPECT_EQ(test1.m_alpha, 255);
+    EXPECT_EQ(test1.m_r, uint8_t(0));
+    EXPECT_EQ(test1.m_g, uint8_t(0));
+    EXPECT_EQ(test1.m_b, uint8_t(255));
+    EXPECT_EQ(test1.m_alpha, uint8_t(255));
 
     m_prop->setVal(gt::rgb(255, 255, 0, 0).toHexString());
     gt::rgb test2 = m_prop->toRGB();
-    EXPECT_EQ(test2.m_r, 255);
-    EXPECT_EQ(test2.m_g, 255);
-    EXPECT_EQ(test2.m_b, 0);
-    EXPECT_EQ(test2.m_alpha, 0);
+    EXPECT_EQ(test2.m_r, uint8_t(255));
+    EXPECT_EQ(test2.m_g, uint8_t(255));
+    EXPECT_EQ(test2.m_b, uint8_t(0));
+    EXPECT_EQ(test2.m_alpha, uint8_t(0));
 }
 
 TEST_F(TestGtColorProperty, setFromRGB)
